melib/backends: replace watch() with watcher(), BackendWatcher trait
```
Return a [`Box<dyn BackendWatcher>`](BackendWatcher), to which you can
register the mailboxes you are interested in for updates and then
consume to spawn a watching `Future`.  The `Future` sends events to the
[`BackendEventConsumer`](BackendEventConsumer) supplied to the backend
in its constructor method.
```

Watching mailboxes for updates is more flexible now that you can
explicitly register mailboxes and set polling period.